DDPB-2635 Added check for keys being the same

+++ src/AppBundle/Service/File/Storage/S3Storage.php
@@ -105,42 +105,44 @@ class S3Storage implements StorageInterface
      */
     public function removeFromS3($key)
     {
-        /*
-         * ListObjectVersions is permitted by ListBucketVersions in IAM.
-         */
-        $objectVersions = $this->s3Client->listObjectVersions([
-            'Bucket' => $this->bucketName,
-            'Prefix' => $key
-        ]);
+        if (!empty($key)) {
+            /*
+             * ListObjectVersions is permitted by ListBucketVersions in IAM.
+             */
+            $objectVersions = $this->s3Client->listObjectVersions([
+                'Bucket' => $this->bucketName,
+                'Prefix' => $key
+            ]);
 
-        if ($objectVersions instanceof \Aws\Result) {
-            if (array_key_exists('Versions',$objectVersions)) {
-                foreach ($objectVersions['Versions'] as $versionData) {
-                    if (!empty($versionData["VersionId"])) {
-                        $this->s3Client->deleteObject([
-                            'Bucket' => $this->bucketName,
-                            'Key' => $versionData['Key'],
-                            'VersionId' => $versionData['VersionId'],
-                        ]);
+            if ($objectVersions instanceof \Aws\Result) {
+                if (array_key_exists('Versions', $objectVersions)) {
+                    foreach ($objectVersions['Versions'] as $versionData) {
+                        if (!empty($versionData["VersionId"]) && $versionData['Key'] == $key) {
+                            $this->s3Client->deleteObject([
+                                'Bucket' => $this->bucketName,
+                                'Key' => $versionData['Key'],
+                                'VersionId' => $versionData['VersionId'],
+                            ]);
+                        }
                     }
                 }
-            }
 
-            if (array_key_exists('DeleteMarkers',$objectVersions)) {
-                // remove any deleteMarkers permanently
-                foreach ($objectVersions['DeleteMarkers'] as $dmData) {
-                    if (!empty($dmData["VersionId"])) {
-                        $this->s3Client->deleteObject([
-                            'Bucket' => $this->bucketName,
-                            'Key' => $dmData['Key'],
-                            'VersionId' => $dmData['VersionId'],
-                        ]);
+                if (array_key_exists('DeleteMarkers', $objectVersions)) {
+                    // remove any deleteMarkers permanently
+                    foreach ($objectVersions['DeleteMarkers'] as $dmData) {
+                        if (!empty($dmData["VersionId"]) && $dmData['Key'] == $key) {
+                            $this->s3Client->deleteObject([
+                                'Bucket' => $this->bucketName,
+                                'Key' => $dmData['Key'],
+                                'VersionId' => $dmData['VersionId'],
+                            ]);
+                        }
                     }
                 }
+
+                return true;
             }
-            return true;
         }
-
         throw new \RuntimeException('Could not remove from S3: Version data not found');
     }
